fix: rename author-headers (#144)
